+++ Source/Core/SectionModel.swift
@@ -57,7 +57,7 @@ open class SectionModel: Section, SupplementaryAccessible
     /// Returns items of `type` in current section
     open func items<T>(ofType type: T.Type) -> [T]
     {
-        return items.flatMap { $0 as? T }
+        return items.compactMap { $0 as? T }
     }
     
     /// Number of items in current section